pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "strands-xai"
version = "0.1.0"
description = "xAI model provider for Strands Agents SDK"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Francesco Cerizzi" }
]
keywords = ["strands", "agents", "xai", "grok", "ai", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "strands-agents>=1.23.0",
    "xai-sdk>=1.5.0,<2.0.0",
]

[project.urls]
Homepage = "https://github.com/Cerrix/strands-xai"
Documentation = "https://github.com/Cerrix/strands-xai#readme"
Repository = "https://github.com/Cerrix/strands-xai"
Issues = "https://github.com/Cerrix/strands-xai/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "twine>=4.0.0",
    "build>=0.10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/strands_xai"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
